Added the OffsetTuple

+++ file/offset_tuple.go
@@ -2,5 +2,5 @@ package file
 
 type OffSetTuple struct {
 	b int64
-	r int32
+	r int64
 }
